soluzion_test_client/requirements.txt
python-socketio[client]~=5.11.2